+++ change-log.md
@@ -2,6 +2,7 @@
 
 - Fixed issue with auto indent not working when keyword case transform was disabled (#24)
 - Fixed issue with REM comment followed by newline, which caused next line to be a comment (#28)
+- Fixed various spelling errors (#26)
 
 # Change since 0.5.0-beta.1