+++ change-log.md
@@ -3,6 +3,7 @@
 - Fixed issue with auto indent not working when keyword case transform was disabled (#24)
 - Fixed issue with REM comment followed by newline, which caused next line to be a comment (#28)
 - Fixed various spelling errors (#26)
+- Fixed bug with console not being properly updated after compile generated output (#32)
 
 # Change since 0.5.0-beta.1