+++ change-log.md
@@ -5,6 +5,7 @@
 - Fixed various spelling errors (#26)
 - Fixed bug with console not being properly updated after compile generated output (#32)
 - Fixed bug with indenting #defines containing IF keyword, which was counted as ppIF (#34)
+- Fixed toolbar tooltip to shohw long help string (#36)
 
 # Change since 0.5.0-beta.1